requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio>=3.4.3
APScheduler>=3.9.1
uuid>=1.30
SQLite3>=3.35.0
